ClipboardPopover from this commit demonstrates...
If you do not have useEffect, with the open brackets, then setInterval goes on forever.  But that does not mean the component is around.  It is destroyed in the parent.  It just means, cleanup was needed.  The [] at the end of useEffect calls the contents of useEffect when the popover is created and destroyed. By returning a function that calls clearInterval inside useEffect the timer is cleaned up.  Together this shows that the popover component is effectively being destroyed.